fix circular import in pyramid.viewderivers
this is a simple fix for 1.9-branch - a more comprehensive fix will be
done on master

fixes #3112
